[TIMOB-17346] Fixed issues with config.json being overwritten at the same time.
Improved handling of corrupt a config.json file.

Fixed backwards compatibility with Titanium SDK 3.3.0 and older when building an iOS app and Xcode 6 or newer is installed.

Added support for a "helpNoPrompt()" callback when a missing option is encountered.

Fixed bug with abbreviated options that don't have a value being set to true.